+++ style/ansi.go
@@ -1,0 +1,11 @@
+package style
+
+const (
+	Reset       = "\033[0m"
+	ANSIBold    = "\033[1m"
+	ANSIFaint   = "\033[2m"
+	Italic      = "\033[3m"
+	Reverse     = "\033[7m"
+	ItalicOff   = "\033[23m"
+	ANSIMagenta = "\033[35m"
+)