+++ CHANGELOG.md
@@ -8,6 +8,7 @@ _WIP_
 - Replace terminal-dimensions with golang.org/x/term
 - Add `u` keybinding to toggle read/unread status on stories
 - Improve error messages for timeouts, non-200 responses, and unexpected server responses
+- Add support for customizing the theme
 
 ## 3.9